Done round of refactoring on coloring
