+++ app/assets/javascripts/reader_profile_january/BenchmarkCohortChart.story.js
@@ -7,7 +7,7 @@ storiesOf('reader_profile_january/BenchmarkCohortChart', module) // eslint-disab
   .add('default', () => {
     mockFetch();
     return (
-      <div style={{width: 300, border: '1px dashed red', margin: 10}}>
+      <div style={{width: 300, margin: 10}}>
         {testRender(testProps())}
       </div>
     );
